Convert HWRP to custom resource
